updated parse.py merger
